renamed package of wear app in attempt to get messages to send. need to figure out  if there is a way to see if these are being truy sent
