+++ FinalProject/.idea/modules.xml
@@ -4,7 +4,7 @@
     <modules>
       <module fileurl="file://$PROJECT_DIR$/FinalProject.iml" filepath="$PROJECT_DIR$/FinalProject.iml" />
       <module fileurl="file://$PROJECT_DIR$/app/app.iml" filepath="$PROJECT_DIR$/app/app.iml" />
-      <module fileurl="file://$PROJECT_DIR$/finalprojectwear/finalprojectwear.iml" filepath="$PROJECT_DIR$/finalprojectwear/finalprojectwear.iml" />
+      <module fileurl="file://$PROJECT_DIR$/finalproject/finalproject.iml" filepath="$PROJECT_DIR$/finalproject/finalproject.iml" />
     </modules>
   </component>
 </project>
