+++ FinalProject/app/src/main/java/com/kalvi_000/finalproject/WearListenerService.java
@@ -1,5 +1,12 @@
 package com.kalvi_000.finalproject;
 
+import android.content.Intent;
+import android.net.Uri;
+import android.util.Log;
+import android.widget.Toast;
+
+import com.google.android.gms.wearable.DataEventBuffer;
+import com.google.android.gms.wearable.MessageEvent;
 import com.google.android.gms.wearable.WearableListenerService;
 
 /**
@@ -7,6 +14,22 @@ import com.google.android.gms.wearable.WearableListenerService;
  */
 
 public class WearListenerService extends WearableListenerService {
+
+    @Override
+    public void onMessageReceived(MessageEvent messageEvent) {
+        showToast(messageEvent.getPath());
+        Log.d("OnMessageRevieved *****", "onMessageRecieved Fired");
+    }
+
+    @Override
+    public void onDataChanged(DataEventBuffer dataEventBuffer) {
+        super.onDataChanged(dataEventBuffer);
+        Log.d("onDataChanged *****", "onDataChanged Fired");
+    }
+
+    private void showToast(String message) {
+        Toast.makeText(this, message, Toast.LENGTH_LONG).show();
+    }
 
 
 }
