+++ FinalProject/finalproject/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.kalvi_000.finalprojectwear" >
+    package="com.kalvi_000.finalproject" >
 
     <uses-feature android:name="android.hardware.type.watch" />
 
@@ -17,7 +17,7 @@
             android:required="false" />
 
         <activity
-            android:name=".MainActivityWear"
+            android:name="com.kalvi_000.finalproject.MainActivityWear"
             android:label="@string/app_name"
             android:theme="@android:style/Theme.DeviceDefault.Light" >
             <intent-filter>
