+++ FinalProject/finalproject/src/main/res/layout/activity_main_wear.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:id="@+id/container"
-    tools:context="com.kalvi_000.finalprojectwear.MainActivityWear"
+    tools:context="com.kalvi_000.finalproject.MainActivityWear"
     tools:deviceIds="wear">
 
     <TextView
