+++ FinalProject/settings.gradle
@@ -1,1 +1,1 @@
-include ':app', ':finalprojectwear'
+include ':app', ':finalproject'
